+++ python/cli/prelude_cli/views/scm.py
@@ -3,8 +3,8 @@ import requests
 from time import sleep
 
 from prelude_cli.views.shared import Spinner, pretty_print
-from prelude_sdk.controllers.detect_controller import DetectController
 from prelude_sdk.controllers.export_controller import ExportController
+from prelude_sdk.controllers.jobs_controller import JobsController
 from prelude_sdk.controllers.scm_controller import ScmController
 from prelude_sdk.models.codes import Control
 
@@ -96,10 +96,9 @@ def sync(controller, partner):
     """ Update policy evaluation for given partner """
     with Spinner(description='Updating policy evaluation'):
         job_id = controller.update_evaluation(partner=Control[partner])['job_id']
-        detect = DetectController(account=controller.account)
-        while (result := detect.get_background_job(job_id))['end_time'] is None:
+        jobs = JobsController(account=controller.account)
+        while (result := jobs.job_status(job_id))['end_time'] is None:
             sleep(3)
-            pass
         return result
     
 @scm.command('export')
@@ -115,12 +114,11 @@ def sync(controller, partner):
 def export(controller, type, output_file, limit, odata_filter, odata_orderby, partner):
     """ Export SCM data """
     with Spinner(description='Exporting SCM data'):
-        detect = DetectController(account=controller.account)
         export = ExportController(account=controller.account)
+        jobs = JobsController(account=controller.account)
         job_id = export.export_scm(export_type=type, filter=odata_filter, orderby=odata_orderby, partner=Control[partner] if partner else None, top=limit)['job_id']
-        while (result := detect.get_background_job(job_id))['end_time'] is None:
+        while (result := jobs.job_status(job_id))['end_time'] is None:
             sleep(3)
-            pass
         if result['successful']:
             data = requests.get(result['results']['url'], timeout=10).content
             with open(output_file, 'wb') as f:
